refactor: Receipt 리포지토리 리팩토링
