+++ src/main/java/org/indoles/receiptserviceserver/core/receipt/infra/ReceiptRepository.java
@@ -1,7 +1,8 @@
-package org.indoles.receiptserviceserver.core.receipt.domain;
+package org.indoles.receiptserviceserver.core.receipt.infra;
 
 
 
+import org.indoles.receiptserviceserver.core.receipt.domain.Receipt;
 import org.indoles.receiptserviceserver.core.receipt.dto.BuyerReceiptSearchCondition;
 import org.indoles.receiptserviceserver.core.receipt.dto.SellerReceiptSearchCondition;
